refactor: 리스트 페이지 제외하고 components폴더로 폴더링
