+++ src/pages/ListPage/components/IdolCard.jsx
